requirements.txt
bs4
cffi
chardet
Click
cryptography
Flask
Flask-Compress
Flask-Cors
Flask-JSON
gast
google-pasta
grpcio
h5py
idna
itsdangerous
Jinja2
Keras-Applications
Keras-Preprocessing
keyring
Markdown
MarkupSafe
mock
numpy
opencv-contrib-python
opencv-python
pdf2image
Pillow
protobuf
pycparser
pyxdg
requests
scipy
SecretStorage
six
soupsieve
tensorboard
tensorflow--gpu==1.13.1
termcolor
torch
urllib3
Werkzeug
tqdm
